requirements.dev.txt
# dev dependencies
ipython>=7.0.0,<=8.0.0
ipdb
pudb
pytest-watch

# Code Quality
flake8
pyproject-flake8
black
isort

# Install project as editable
-e .s